backend/requirements.txt
django==3.2.16
djangorestframework==3.12.4
pillow==9.0.0
django-annoying==0.10.7
django-filter==23.5
psycopg2-binary==2.9.3
djoser==2.2.3